# feat: improve content type read performance