--- a/lib/validation.js
+++ b/lib/validation.js
@@ -261,7 +261,74 @@ function wrapValidationError (result, dataVar, schemaErrorFormatter) {
  */
 function getEssenceMediaType (header) {
   if (!header) return ''
-  return header.split(/[ ;]/, 1)[0].trim().toLowerCase()
+
+  // The Content-Type header field indicates the media type of the resource
+  // The media type format is:
+  // media-type = type "/" subtype parameters
+  // type       = token
+  // subtype    = token
+  // The type/subtype MAY be followed by semicolon-delimited parameters
+  // https://datatracker.ietf.org/doc/html/rfc9110#section-8.3.1
+
+  // Tokens are case-insensitive and consist of:
+  // https://datatracker.ietf.org/doc/html/rfc9110#name-tokens
+
+  //   tchar          = "!" / "#" / "$" / "%" / "&" / "'" / "*"
+  //                  / "+" / "-" / "." / "^" / "_" / "`" / "|" / "~"
+  //                  / DIGIT / ALPHA
+  //                  ; any VCHAR, except delimiters
+  // Many HTTP field values are defined using common syntax components, separated by whitespace or specific delimiting characters.
+  // Delimiters are chosen from the set of US-ASCII visual characters not allowed in a token (DQUOTE and "(),/:;<=>?@[\]{}").
+
+  header = header.trim().toLowerCase()
+  let firstToken = true
+  for (let i = 0; i < header.length; i++) {
+    const char = header.codePointAt(i)
+    const isDelimiterChar = isDelimiter(char)
+    if (!isDelimiterChar) {
+      continue
+    }
+
+    if (firstToken) {
+      if (char !== 47) {
+        // The first delimiter MUST be a slash
+        throw new Error('Invalid Content-Type header')
+      } else {
+        firstToken = false
+      }
+    } else { // Look for the second delimiter
+      if (char !== 59 && char !== 32) {
+        // The second delimiter MUST be a semicolon or a space
+        throw new Error('Invalid Content-Type header')
+      }
+      return header.slice(0, i)
+    }
+  }
+
+  // If we reach here, it means we didn't find a semicolon
+  return header
+}
+
+function isDelimiter (charCode) {
+  // Delimiters are defined by the RFC 9110: (DQUOTE and "(),/:;<=>?@[\]{}")
+  return charCode <= 32 || // NOT any VCHAR (visible char): these must be not included in token
+    charCode === 34 || // DQUOTE
+    charCode === 40 || // (
+    charCode === 41 || // )
+    charCode === 44 || // ,
+    charCode === 47 || // /
+    charCode === 58 || // :
+    charCode === 59 || // ;
+    charCode === 60 || // <
+    charCode === 61 || // =
+    charCode === 62 || // >
+    charCode === 63 || // ?
+    charCode === 64 || // @
+    charCode === 91 || // [
+    charCode === 92 || // \
+    charCode === 93 || // ]
+    charCode === 123 || // {
+    charCode === 125 // }
 }
 
 module.exports = {