--- a/lib/validation.js
+++ b/lib/validation.js
@@ -281,54 +281,24 @@ function getEssenceMediaType (header) {
   // Delimiters are chosen from the set of US-ASCII visual characters not allowed in a token (DQUOTE and "(),/:;<=>?@[\]{}").
 
   header = header.trim().toLowerCase()
-  let firstToken = true
-  for (let i = 0; i < header.length; i++) {
-    const char = header.codePointAt(i)
-    const isDelimiterChar = isDelimiter(char)
-    if (!isDelimiterChar) {
-      continue
-    }
+  let slashIndex = -1
 
-    if (firstToken) {
-      if (char !== 47) {
-        // The first delimiter MUST be a slash
-        throw new Error('Invalid Content-Type header')
-      } else {
-        firstToken = false
+  const length = header.length
+  for (let i = 0; i < length; i++) {
+    const char = header[i]
+
+    if (char === '/') { // Slash ('/')
+      if (slashIndex === -1) {
+        slashIndex = i
       }
-    } else { // Look for the second delimiter
-      if (char !== 59 && char !== 32) {
-        // The second delimiter MUST be a semicolon or a space
-        throw new Error('Invalid Content-Type header')
-      }
+    } else if (char === ';' || char === ' ') { // Semicolon (';') or Space (' ')
       return header.slice(0, i)
+    } else if (slashIndex !== -1 && char <= ' ') { // Invalid character after slash
+      throw new Error('Invalid Content-Type header')
     }
   }
 
-  // If we reach here, it means we didn't find a semicolon
   return header
-}
-
-function isDelimiter (charCode) {
-  // Delimiters are defined by the RFC 9110: (DQUOTE and "(),/:;<=>?@[\]{}")
-  return charCode <= 32 || // NOT any VCHAR (visible char): these must be not included in token
-    charCode === 34 || // DQUOTE
-    charCode === 40 || // (
-    charCode === 41 || // )
-    charCode === 44 || // ,
-    charCode === 47 || // /
-    charCode === 58 || // :
-    charCode === 59 || // ;
-    charCode === 60 || // <
-    charCode === 61 || // =
-    charCode === 62 || // >
-    charCode === 63 || // ?
-    charCode === 64 || // @
-    charCode === 91 || // [
-    charCode === 92 || // \
-    charCode === 93 || // ]
-    charCode === 123 || // {
-    charCode === 125 // }
 }
 
 module.exports = {