@jsumners 2025-04-25
I spent some time yesterday adding my suggestion, but ended up deleting the branch. The code in question is only used when a schema is defined that tries to verify content types. The baseline code for every incoming request merely checks that the header is set and has a value of something other than an empty string. So I decided it wasn't worth it.