@Eomm 2025-04-25
I think we can propagate the code to the content-type parser as well - the target here was to be compliant with the RFC compliant first, avoid any potential bad actors and (luckily) improve the performance